refactor(gateway): split server.rs into focused submodules
server.rs (2135 lines) converted to a directory-based module with 14
focused files. No public API changes — lib.rs re-exports DaemonServer
and DaemonStartOptions unchanged.

New structure:
  server/mod.rs        — DaemonServer struct, SessionHandle, accessors
  server/paths.rs      — DaemonPaths struct and all path methods
  server/startup.rs    — DaemonStartOptions + DaemonServer::start()
  server/lifecycle.rs  — shutdown_plugins, shutdown_servers, cleanup
  server/monitoring.rs — AbortOnDrop, spawn_health/ephemeral/cleanup loops
  server/plugins.rs    — WatcherReloadContext, plugin watcher, hot-reload
  server/rpc/mod.rs    — RpcImpl struct + thin AstridRpcServer delegation
  server/rpc/workspace.rs  — ws_ns, workspace KV helpers
  server/rpc/session.rs    — session CRUD RPC impl
  server/rpc/approval.rs   — approval/elicitation/cancel impl
  server/rpc/mcp_servers.rs — MCP server RPC impl
  server/rpc/plugins.rs    — plugin list/load/unload impl
  server/rpc/budget.rs     — budget/allowances/audit impl
  server/rpc/events.rs     — subscribe_events/shutdown/list_tools impl

Verified: cargo check, cargo test (63 tests pass), cargo clippy -D warnings
