+++ crates/astrid-gateway/src/server/startup.rs
@@ -1,0 +1,390 @@
+//! Daemon startup logic: configuration loading, component initialization, server binding.
+
+use std::collections::{HashMap, HashSet};
+use std::net::SocketAddr;
+use std::path::PathBuf;
+use std::sync::Arc;
+use std::sync::atomic::AtomicUsize;
+use std::time::{Duration, Instant};
+
+use astrid_approval::budget::{WorkspaceBudgetSnapshot, WorkspaceBudgetTracker};
+use astrid_audit::AuditLog;
+use astrid_capabilities::CapabilityStore;
+use astrid_crypto::KeyPair;
+use astrid_hooks::{HookManager, discover_hooks};
+use astrid_llm::{ClaudeProvider, LlmProvider, OpenAiCompatProvider, ZaiProvider};
+use astrid_mcp::McpClient;
+use astrid_plugins::manifest::PluginEntryPoint;
+use astrid_plugins::{
+    PluginContext, PluginId, PluginRegistry, WasmPluginLoader, discover_manifests,
+};
+use astrid_runtime::{AgentRuntime, SessionStore, config_bridge};
+use astrid_storage::{KvStore, ScopedKvStore, SurrealKvStore};
+use jsonrpsee::server::{Server, ServerHandle};
+use tokio::sync::{RwLock, broadcast};
+use tracing::{info, warn};
+
+use super::paths::DaemonPaths;
+use super::rpc::RpcImpl;
+use super::{DaemonServer, SessionHandle};
+use crate::rpc::AstridRpcServer;
+
+/// Options controlling daemon startup behaviour.
+#[derive(Debug, Clone, Default)]
+pub struct DaemonStartOptions {
+    /// When `true`, the daemon shuts down automatically after all clients
+    /// disconnect and the grace period elapses.
+    pub ephemeral: bool,
+    /// Override for the idle-shutdown grace period (seconds). Falls back to
+    /// `gateway.idle_shutdown_secs` from the config.
+    pub grace_period_secs: Option<u64>,
+}
+
+impl DaemonServer {
+    /// Create and start a new daemon server.
+    ///
+    /// Binds to `127.0.0.1:0` (OS picks a free port), writes port/PID files,
+    /// and returns the server handle for lifecycle management.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the server cannot bind or required components fail to initialize.
+    #[allow(clippy::too_many_lines)]
+    pub async fn start(
+        options: DaemonStartOptions,
+    ) -> Result<(Self, ServerHandle, SocketAddr, astrid_config::Config), crate::GatewayError> {
+        let paths = DaemonPaths::default_dir().map_err(|e| {
+            crate::GatewayError::Runtime(format!("Failed to resolve daemon paths: {e}"))
+        })?;
+
+        // Resolve and ensure directory structures.
+        let home = astrid_core::dirs::AstridHome::resolve().map_err(|e| {
+            crate::GatewayError::Runtime(format!("Failed to resolve home directory: {e}"))
+        })?;
+        home.ensure().map_err(|e| {
+            crate::GatewayError::Runtime(format!(
+                "Failed to create home directory {}: {e}",
+                home.root().display()
+            ))
+        })?;
+
+        let cwd = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
+        let ws = astrid_core::dirs::WorkspaceDir::detect(&cwd);
+        // Ensure workspace dir and generate workspace ID (idempotent).
+        ws.ensure().map_err(|e| {
+            crate::GatewayError::Runtime(format!(
+                "Failed to ensure workspace directory {}: {e}",
+                ws.root().display()
+            ))
+        })?;
+        let workspace_id = ws.workspace_id().map_err(|e| {
+            crate::GatewayError::Runtime(format!("Failed to get workspace ID: {e}"))
+        })?;
+
+        // Load unified configuration.
+        let cfg = match astrid_config::Config::load(Some(&cwd)) {
+            Ok(r) => r.config,
+            Err(e) => {
+                warn!(error = %e, "Failed to load config; falling back to defaults");
+                astrid_config::Config::default()
+            },
+        };
+
+        // Create LLM provider via config bridge (api key comes from config
+        // precedence chain -- no redundant env var read needed here).
+        // If no key is configured yet, the provider starts with an empty key
+        // and returns a clear error on the first actual LLM call.
+        let provider_config = config_bridge::to_provider_config(&cfg);
+        let llm: Box<dyn LlmProvider> = match cfg.model.provider.as_str() {
+            "zai" => Box::new(ZaiProvider::new(provider_config)),
+            "openai" | "openai-compat" => {
+                let mut p = OpenAiCompatProvider::custom(
+                    provider_config
+                        .base_url
+                        .as_deref()
+                        .unwrap_or("https://api.openai.com/v1/chat/completions"),
+                    Some(&provider_config.api_key),
+                    &provider_config.model,
+                )
+                .with_max_tokens(provider_config.max_tokens)
+                .with_temperature(provider_config.temperature);
+                if let Some(ctx) = provider_config.context_window {
+                    p = p.with_max_context(ctx);
+                }
+                Box::new(p)
+            },
+            // Default to Claude.
+            _ => Box::new(ClaudeProvider::new(provider_config)),
+        };
+
+        // Load MCP server definitions from unified config.
+        let servers_config = config_bridge::to_servers_config(&cfg);
+        let mcp = McpClient::with_config(servers_config);
+
+        // Auto-connect configured auto-start servers.
+        match mcp.connect_auto_servers().await {
+            Ok(n) if n > 0 => info!(count = n, "Auto-connected MCP servers"),
+            Ok(_) => info!("No auto-start MCP servers configured"),
+            Err(e) => warn!(error = %e, "Error during MCP auto-connect"),
+        }
+
+        // Load key once, get two independent instances (avoids double disk read
+        // and double un-zeroized intermediate buffer).
+        let (audit_key, key) =
+            KeyPair::load_or_generate_pair(home.user_key_path()).map_err(|e| {
+                crate::GatewayError::Runtime(format!("Failed to load/generate key: {e}"))
+            })?;
+
+        let audit = AuditLog::open(home.audit_db_path(), audit_key)
+            .map_err(|e| crate::GatewayError::Runtime(format!("Failed to open audit log: {e}")))?;
+
+        let sessions = SessionStore::from_home(&home);
+
+        // Convert workspace and runtime config via bridge.
+        let config = config_bridge::to_runtime_config(&cfg, &cwd);
+
+        let hook_manager = HookManager::new();
+        let hooks_extra = vec![home.hooks_dir()];
+        let discovered = discover_hooks(Some(&hooks_extra));
+        hook_manager.register_all(discovered).await;
+
+        // Clone MCP client for plugin registry before moving into runtime.
+        let mcp_for_plugins = mcp.clone();
+        // Clone MCP client for watcher-driven plugin reloads.
+        let mcp_for_watcher = mcp.clone();
+
+        // Open persistent capability store (tokens survive restarts).
+        let capabilities_store = Arc::new(
+            CapabilityStore::with_persistence(home.capabilities_db_path()).map_err(|e| {
+                crate::GatewayError::Runtime(format!("Failed to open capabilities store: {e}"))
+            })?,
+        );
+
+        // Open workspace state KV store (allowances, budget, escape state).
+        // Lives in ~/.astrid/state.db, namespaced by workspace ID.
+        let workspace_kv: Arc<dyn KvStore> =
+            Arc::new(SurrealKvStore::open(home.state_db_path()).map_err(|e| {
+                crate::GatewayError::Runtime(format!("Failed to open workspace state store: {e}"))
+            })?);
+
+        // Open persistent KV store for deferred resolution queue.
+        let deferred_kv: Arc<dyn KvStore> =
+            Arc::new(SurrealKvStore::open(home.deferred_db_path()).map_err(|e| {
+                crate::GatewayError::Runtime(format!("Failed to open deferred store: {e}"))
+            })?);
+
+        // Discover and register plugins (does not load them yet).
+        // Built before the runtime so we can pass the Arc into new_arc().
+        let mut plugin_registry = PluginRegistry::new();
+        let wasm_loader = Arc::new(WasmPluginLoader::new());
+        let plugin_dirs = vec![home.plugins_dir()];
+        let discovered = discover_manifests(Some(&plugin_dirs));
+        for (mut manifest, plugin_dir) in discovered {
+            // Resolve relative WASM paths to absolute, anchored at the
+            // directory where the manifest was discovered. Without this,
+            // WasmPlugin::do_load would resolve them against the daemon CWD
+            // which is wrong for both user-level and workspace-level plugins.
+            if let PluginEntryPoint::Wasm { ref mut path, .. } = manifest.entry_point
+                && path.is_relative()
+            {
+                *path = plugin_dir.join(&*path);
+            }
+
+            let plugin: Box<dyn astrid_plugins::Plugin> = match &manifest.entry_point {
+                PluginEntryPoint::Wasm { .. } => {
+                    Box::new(wasm_loader.create_plugin(manifest.clone()))
+                },
+                PluginEntryPoint::Mcp { .. } => {
+                    match astrid_plugins::create_plugin(
+                        manifest.clone(),
+                        Some(mcp_for_plugins.clone()),
+                        Some(plugin_dir.clone()),
+                    ) {
+                        Ok(p) => p,
+                        Err(e) => {
+                            warn!(plugin = %manifest.id, error = %e, "Failed to create MCP plugin");
+                            continue;
+                        },
+                    }
+                },
+            };
+            if let Err(e) = plugin_registry.register(plugin) {
+                warn!(plugin = %manifest.id, error = %e, "Failed to register plugin");
+            }
+        }
+        let plugin_registry = Arc::new(RwLock::new(plugin_registry));
+
+        let runtime = AgentRuntime::new_arc(
+            llm,
+            mcp,
+            audit,
+            sessions,
+            key,
+            config,
+            Some(hook_manager),
+            Some(Arc::clone(&plugin_registry)),
+        );
+
+        let (shutdown_tx, _) = broadcast::channel(1);
+        let session_map: Arc<RwLock<HashMap<astrid_core::SessionId, SessionHandle>>> =
+            Arc::new(RwLock::new(HashMap::new()));
+
+        // Build and start the jsonrpsee server.
+        let server = Server::builder()
+            .build("127.0.0.1:0")
+            .await
+            .map_err(|e| crate::GatewayError::Runtime(format!("Failed to bind server: {e}")))?;
+
+        let addr = server
+            .local_addr()
+            .map_err(|e| crate::GatewayError::Runtime(format!("Failed to get address: {e}")))?;
+
+        // Load workspace budget from KV store and construct the tracker.
+        let ws_max_usd = config_bridge::workspace_max_usd(&cfg);
+        let ws_warn_pct = cfg.budget.warn_at_percent;
+        let ns_budget = super::rpc::workspace::ws_ns(&workspace_id, "budget");
+        let workspace_budget_tracker = match workspace_kv.get(&ns_budget, "all").await {
+            Ok(Some(data)) => {
+                if let Ok(snapshot) = serde_json::from_slice::<WorkspaceBudgetSnapshot>(&data) {
+                    Arc::new(WorkspaceBudgetTracker::restore(
+                        &snapshot,
+                        ws_max_usd,
+                        ws_warn_pct,
+                    ))
+                } else {
+                    Arc::new(WorkspaceBudgetTracker::new(ws_max_usd, ws_warn_pct))
+                }
+            },
+            _ => Arc::new(WorkspaceBudgetTracker::new(ws_max_usd, ws_warn_pct)),
+        };
+
+        let model_name = cfg.model.model.clone();
+        let active_connections = Arc::new(AtomicUsize::new(0));
+        let user_unloaded_plugins: Arc<RwLock<HashSet<PluginId>>> =
+            Arc::new(RwLock::new(HashSet::new()));
+
+        let rpc_impl = RpcImpl {
+            runtime: Arc::clone(&runtime),
+            sessions: Arc::clone(&session_map),
+            plugin_registry: Arc::clone(&plugin_registry),
+            deferred_kv,
+            capabilities_store,
+            workspace_kv: Arc::clone(&workspace_kv),
+            workspace_budget_tracker,
+            started_at: Instant::now(),
+            shutdown_tx: shutdown_tx.clone(),
+            workspace_id,
+            model_name,
+            active_connections: Arc::clone(&active_connections),
+            ephemeral: options.ephemeral,
+            user_unloaded_plugins: Arc::clone(&user_unloaded_plugins),
+            workspace_root: cwd.clone(),
+        };
+
+        let handle = server.start(rpc_impl.into_rpc());
+
+        // Write PID and port files.
+        let pid = std::process::id();
+        std::fs::write(paths.pid_file(), pid.to_string())
+            .map_err(|e| crate::GatewayError::Runtime(format!("Failed to write PID file: {e}")))?;
+        std::fs::write(paths.port_file(), addr.port().to_string())
+            .map_err(|e| crate::GatewayError::Runtime(format!("Failed to write port file: {e}")))?;
+
+        info!(addr = %addr, pid = pid, "Daemon server started");
+
+        // Background task: auto-load discovered plugins after the server is
+        // accepting connections (avoids blocking the 5s CLI connect timeout).
+        // Each plugin is taken out of the registry, loaded without the lock
+        // held, then put back -- so MCP handshakes don't block other operations.
+        {
+            let registry_clone = Arc::clone(&plugin_registry);
+            let kv_clone = Arc::clone(&workspace_kv);
+            let workspace_root = cwd.clone();
+            tokio::spawn(async move {
+                // Collect IDs under a brief read lock.
+                let plugin_ids: Vec<PluginId> = {
+                    let registry = registry_clone.read().await;
+                    registry.list().into_iter().cloned().collect()
+                };
+
+                for plugin_id in plugin_ids {
+                    // Take the plugin out (brief write lock).
+                    let mut plugin = {
+                        let mut registry = registry_clone.write().await;
+                        match registry.unregister(&plugin_id) {
+                            Ok(p) => p,
+                            Err(_) => continue,
+                        }
+                    };
+
+                    let kv = match ScopedKvStore::new(
+                        Arc::clone(&kv_clone),
+                        format!("plugin:{plugin_id}"),
+                    ) {
+                        Ok(kv) => kv,
+                        Err(e) => {
+                            warn!(plugin_id = %plugin_id, error = %e, "Failed to create plugin KV scope");
+                            // Put it back.
+                            let mut registry = registry_clone.write().await;
+                            let _ = registry.register(plugin);
+                            continue;
+                        },
+                    };
+                    let config = plugin.manifest().config.clone();
+                    let ctx = PluginContext::new(workspace_root.clone(), kv, config);
+
+                    // Load without holding any lock.
+                    if let Err(e) = plugin.load(&ctx).await {
+                        warn!(plugin_id = %plugin_id, error = %e, "Failed to auto-load plugin");
+                    } else {
+                        info!(plugin_id = %plugin_id, "Auto-loaded plugin");
+                    }
+
+                    // Put the plugin back (brief write lock).
+                    let mut registry = registry_clone.write().await;
+                    let _ = registry.register(plugin);
+                }
+            });
+        }
+
+        // Floor health interval at 5s to prevent zero/tiny intervals.
+        let health_interval = Duration::from_secs(cfg.gateway.health_interval_secs.max(5));
+
+        // Resolve ephemeral grace period: CLI flag -> config default.
+        let ephemeral_grace_secs = options
+            .grace_period_secs
+            .unwrap_or(cfg.gateway.idle_shutdown_secs);
+
+        let session_cleanup_interval =
+            Duration::from_secs(cfg.gateway.session_cleanup_interval_secs.max(10));
+
+        // Write mode file so callers (e.g. `daemon status`) can read it.
+        let mode_str = if options.ephemeral {
+            "ephemeral"
+        } else {
+            "persistent"
+        };
+        let _ = std::fs::write(paths.mode_file(), mode_str);
+
+        let daemon = Self {
+            runtime,
+            sessions: session_map,
+            plugin_registry,
+            workspace_kv,
+            mcp_client: mcp_for_watcher,
+            wasm_loader,
+            home: home.clone(),
+            workspace_root: cwd,
+            started_at: Instant::now(),
+            shutdown_tx,
+            paths,
+            health_interval,
+            ephemeral: options.ephemeral,
+            ephemeral_grace_secs,
+            active_connections,
+            session_cleanup_interval,
+            user_unloaded_plugins,
+        };
+        Ok((daemon, handle, addr, cfg))
+    }
+}
